requirements.txt
appnope==0.1.0
apted==1.0.3
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
decorator==4.4.1
deepdiff==4.0.8
et-xmlfile==1.0.1
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
jdcal==1.3
jedi==0.15.1
jsonpickle==1.2
jupyter-client==5.3.4
jupyter-core==4.6.1
more-itertools==7.2.0
openpyxl==2.4.0
ordered-set==3.1.1
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
prompt-toolkit==2.0.10
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.2
pytest==5.2.2
python-dateutil==2.8.0
pyzmq==18.1.0
six==1.12.0
tornado==6.0.3
traitlets==4.3.3
vboxapi==1.0
wcwidth==0.1.7
zipp==0.6.0
pyqt==5.9.2
qt==5.9.7
